accel periodic function

+++ paparazzi-simulator/src/main/java/juav/simulator/tasks/sensors/readings/AccelerometerReading.java
@@ -8,8 +8,8 @@ import ub.juav.airborne.math.structs.algebra.Vect3;
  */
 public class AccelerometerReading {
     Vect3<Double> value;
-    int min;
-    int max;
+    double min;
+    double max;
     Mat33<Double> sensitivity;
     Vect3<Double>  neutral;
     Vect3<Double>  noise_std_dev;
@@ -25,19 +25,19 @@ public class AccelerometerReading {
         this.value = value;
     }
 
-    public int getMin() {
+    public double getMin() {
         return min;
     }
 
-    public void setMin(int min) {
+    public void setMin(double min) {
         this.min = min;
     }
 
-    public int getMax() {
+    public double getMax() {
         return max;
     }
 
-    public void setMax(int max) {
+    public void setMax(double max) {
         this.max = max;
     }
